refactoring of login and signup
