+++ backend/java-backend/src/main/java/com/backend/java_backend/Controllers/Authentication/AuthController.java
@@ -31,7 +31,7 @@ public class AuthController {
     @Autowired
     private JwtUtils jwtUtils;
 
-    @PostMapping("/signup")
+    @PostMapping(value = "/signup")
     public ResponseEntity<String> signup( @RequestBody SignupRequest signupRequest) {
         if (userRepository.existsByUsername(signupRequest.getUsername())) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Username is already in use");
@@ -42,9 +42,11 @@ public class AuthController {
         user.setPass(passwordEncoder.encode(signupRequest.getPassword()));
         user.setEmail(signupRequest.getEmail());
         user.setRole(signupRequest.getRole());
+        user.setAddress(signupRequest.getAddress());
+        user.setPhone(signupRequest.getPhoneNumber());
 
         userRepository.save(user);
-        return ResponseEntity.status(HttpStatus.CREATED).body("User registered successfully");
+        return ResponseEntity.status(HttpStatus.CREATED).body("User registered successfully" + user);
     }
 
     @PostMapping(value = "/login", produces = "application/json")
