+++ backend/java-backend/src/main/java/com/backend/java_backend/Controllers/Distributor/DistributorController.java
@@ -1,10 +1,17 @@
 package com.backend.java_backend.Controllers.Distributor;
 
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/distributor")
 public class DistributorController {
+    @GetMapping("/get-info")
+    public ResponseEntity<?> getInfo(){
+        return ResponseEntity.status(HttpStatus.OK).body("DISTRIBUTOR");
+    }
 }
