+++ backend/java-backend/src/main/java/com/backend/java_backend/DTOs/SignupRequest.java
@@ -2,42 +2,15 @@ package com.backend.java_backend.DTOs;
 
 
 import com.backend.java_backend.Classes.User;
+import lombok.Data;
 
+@Data
 public class SignupRequest {
     private String username;
     private String email;
     private String password;
+    private String address;
+    private String phoneNumber;
     private User.Role role;
 
-    public String getEmail() {
-        return email;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public User.Role getRole() {
-        return role;
-    }
-
-    public void setRole(User.Role role) {
-        this.role = role;
-    }
 }
